Fix dataset loading for NTU format and enable multiprocessing for inference

+++ stgcnpp/utils.py
@@ -17,7 +17,7 @@ def _angle_between(v1, v2):
 def _rotation_matrix(axis, theta):
     """Rodrigues rotation (quaternion form) — identical to pyskl."""
     if np.abs(axis).sum() < 1e-6 or np.abs(theta) < 1e-6:
-        return np.eye(3)
+        return np.eye(3, dtype=np.float32)
     axis = axis / np.sqrt(np.dot(axis, axis))
     a = np.cos(theta / 2.0)
     b, c, d = -axis * np.sin(theta / 2.0)
@@ -28,5 +28,6 @@ def _rotation_matrix(axis, theta):
             [aa + bb - cc - dd, 2 * (bc + ad), 2 * (bd - ac)],
             [2 * (bc - ad), aa + cc - bb - dd, 2 * (cd + ab)],
             [2 * (bd + ac), 2 * (cd - ab), aa + dd - bb - cc],
-        ]
+        ],
+        dtype=np.float32,
     )
